requirements.txt
tweepy==4.14.0
python-telegram-bot==20.8
python-dotenv
requests
